Old/job-search-agent/requirements.txt
anthropic>=0.18.0
beautifulsoup4>=4.12.0
requests>=2.31.0
selenium>=4.15.0
webdriver-manager>=4.0.0
python-docx>=1.1.0
schedule>=1.2.0
aiohttp>=3.9.0
jinja2>=3.1.0
python-dotenv>=1.0.0
lxml>=4.9.0
